Add the project-specific config for `gitgitgadget/git` & friends
The idea to let GitGitGadget support projects other than Git is a really
old one, and we're _just_ getting there.

To that end, let's convert the `defaultConfig` (that is highly specific
to the Git-specific support) as of 7cf23ab9 (Merge pull request #1994
from gitgitgadget/use-more-generic-name-for-pr-repo-clone, 2025-09-02)
to JSON and commit it to the `config` branch in `gitgitgadget-workflows`
(which is ultimately where this config is _required_).

Helped-by: Chris. Webster <chris@webstech.net>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>